Breaks out ace_medical_treatment
Moves the ace_medical_treatment section to its own file "ACE_Medical_Treatment.hpp"

Added new file to includes section under config.cpp
